Added pow testcase, fixed some compiler warnings
git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@301 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ read.c
@@ -275,7 +275,7 @@ Boolean readPARM(Ptr p,PARM_T *pparm,char **reasonstr,int fromwin){
 	}
 
 	for(i = 0; i < 8; ++i)
-		slider[i] = pparm->val[i];
+		slider[i] = (uint8_t)pparm->val[i];
 
 	return true;
 }
